+++ src/descriptions/id-broker-api.php
@@ -276,6 +276,23 @@
                 ],
             ],
         ],
+        'assessPasswordInternal' => [
+            'httpMethod' => 'PUT',
+            'uri' => '/user/{employee_id}/password/assess',
+            'responseModel' => 'Result',
+            'parameters' => [
+                'employee_id' => [
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'uri',
+                ],
+                'password' => [
+                    'required' => true,
+                    'type' => 'string',
+                    'location' => 'json',
+                ],
+            ],
+        ],
         'updateUserInternal' => [
             'httpMethod' => 'PUT',
             'uri' => '/user/{employee_id}',